+++ packages/tinacms/src/components/CreateContent.tsx
@@ -38,12 +38,10 @@ import {
   shadow,
 } from '@tinacms/styles'
 import { Dismissible } from 'react-dismissible'
-import { useFrameContext } from './SyledFrame'
 import { useCMS, useSubscribable } from '../react-tinacms'
 
 export const CreateContentMenu = () => {
   const cms = useCMS()
-  const frame = useFrameContext()
   const [visible, setVisible] = React.useState(false)
 
   const contentCreatorPlugins = cms.plugins.findOrCreateMap('content-creator')
@@ -61,7 +59,6 @@ export const CreateContentMenu = () => {
             click
             escape
             onDismiss={() => setVisible(false)}
-            document={frame.document}
             disabled={!visible}
           >
             {contentCreatorPlugins.all().map(plugin => (
@@ -149,7 +146,7 @@ const ContentMenuWrapper = styled.div`
 `
 
 const ContentMenu = styled.div<{ open: boolean }>`
-  min-width: 12rem;
+  min-width: 192px;
   border-radius: ${radius()};
   border: 1px solid ${color.grey(2)};
   display: block;
@@ -171,7 +168,7 @@ const ContentMenu = styled.div<{ open: boolean }>`
     css`
       opacity: 1;
       pointer-events: all;
-      transform: translate3d(0, 2.75rem, 0) scale3d(1, 1, 1);
+      transform: translate3d(0, 44px, 0) scale3d(1, 1, 1);
     `};
 `
 
@@ -179,8 +176,8 @@ const CreateButton = styled.button`
   position: relative;
   text-align: center;
   font-size: ${font.size(0)};
-  padding: 0 0.75rem;
-  height: 2.5rem;
+  padding: 0 12px;
+  height: 40px;
   font-weight: 500;
   width: 100%;
   background: none;